// frontend/src/api/models/catalogAttachmentListResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export interface CatalogAttachmentListResponse {
  content_type?: string
  created_at?: string
  file_name?: string
  file_size?: number
  id?: string
  product_id?: string
  sort_order?: number
  status?: string
  thumbnail_url?: string
  type?: string
  uploaded_by?: string
  url?: string
}